flattens parser directory
