+++ src/main.rs
@@ -29,6 +29,7 @@ For more information see the README.\n
     );
 
     let tokens: Vec<lexer::Token> = lexer::lex(&cli_args.source);
+	
     let ast: parser::ast::AST = parser::parse(tokens);
     let interpreter = interpreter::run(ast, cli_args.dir, cli_args.recursive);
     block_on(interpreter); // wait interpreter to finish
