+++ src/parser/conditionals.rs
@@ -1,7 +1,9 @@
 use super::*;
-pub use conditionals::Conditionals;
 
-pub mod conditionals;
+pub trait Conditionals {
+    fn parse_next_conditional(&mut self) -> super::Conditional;
+    fn reset_conditionals(&mut self);
+}
 
 impl Conditionals for super::Parser {
     fn parse_next_conditional(&mut self) -> Conditional {
