--- src/parser/conditionals/conditionals.rs
@@ -1,4 +1,0 @@
-pub trait Conditionals {
-    fn parse_next_conditional(&mut self) -> super::Conditional;
-    fn reset_conditionals(&mut self);
-}
