+++ src/parser/cycles.rs
@@ -1,7 +1,10 @@
 use super::*;
-pub use cycles::Cycles;
 
-pub mod cycles;
+pub trait Cycles {
+    fn push_current_cycle(&mut self, tokens_index_list: &mut Vec<usize>);
+    fn parse_raw_cycles(&mut self);
+    fn parse_cycles(&mut self);
+}
 
 impl Cycles for super::Parser {
     // i do not remember well what does this, but works
