--- src/parser/cycles/cycles.rs
@@ -1,5 +1,0 @@
-pub trait Cycles {
-    fn push_current_cycle(&mut self, tokens_index_list: &mut Vec<usize>);
-    fn parse_raw_cycles(&mut self);
-    fn parse_cycles(&mut self);
-}
