--- src/parser/exceptions/exceptions.rs
@@ -1,3 +1,0 @@
-pub trait Exceptions {
-    fn raise_parse_error(&self, err_type: &str, msg: String);
-}
