--- src/parser/operations/operations.rs
@@ -1,3 +1,0 @@
-pub trait Operations {
-    fn parse_operations(&mut self) -> Vec<super::Cycle>;
-}
